path argument must be absolute or will segfault
If you give a relative path argument to the constructor, the program
will segfault at FSEventStreamCreate().

Since I think it's better to use XS as little as possible, I wrote the
check in the Perl code and moved the underlying XS constructor to a
private method.

An alternative could be to use Cwd to ensure all paths are absolute, but
that's up for discussion. This at least prevents very terrible things
from happening.