Refer to a copy of the respipe in watch()'s return
The way this used was that we'd return a glob value that referred
to the file descriptor for the results pipe we use to indicate new
values from FSWatch.  The problem was that if you call stop(), that
would close the pipe, invalidating the file descriptor, and a separate
issue would cause that blob to hang around until global destruction
time.

Now we just dup() the file descriptor so that the glob value we return
has its own file descriptor to close